refactor: define scanners & searchers in index
