+++ src/scanner/index.ts
@@ -1,5 +1,18 @@
+import type { Scanner } from "@/types";
+
+import { Browserling } from "./browserling";
+import { HybridAnalysis } from "./hybridanalysis";
+import { URLScan } from "./urlscan";
+import { VirusTotal } from "./virustotal";
+
 export { Browserling } from "./browserling";
 export { HybridAnalysis } from "./hybridanalysis";
-export { Scanners } from "./scanners";
 export { URLScan } from "./urlscan";
 export { VirusTotal } from "./virustotal";
+
+export const Scanners: Scanner[] = [
+  new Browserling(),
+  new HybridAnalysis(),
+  new URLScan(),
+  new VirusTotal(),
+];
