+++ src/searcher/index.ts
@@ -1,3 +1,74 @@
+import type { Searcher } from "@/types";
+
+import {
+  APKLab,
+  AbuseIPDB,
+  AnyRun,
+  ArchiveOrg,
+  ArchiveToday,
+  BGPView,
+  BinaryEdge,
+  BitcoinAbuse,
+  BitcoinWhosWho,
+  BlockChain,
+  BlockCypher,
+  Blockchair,
+  Censys,
+  CheckPhish,
+  Crtsh,
+  DNSlytics,
+  DomainTools,
+  DomainWatch,
+  EmailRep,
+  ExploitDatabase,
+  FileScan,
+  FortiGuard,
+  GoogleSafeBrowsing,
+  GreyNoise,
+  Hashdd,
+  HurricaneElectric,
+  HybridAnalysis,
+  IPIP,
+  IPinfo,
+  InQuest,
+  IntelligenceX,
+  Intezer,
+  JoeSandbox,
+  Malshare,
+  Maltiverse,
+  MalwareBazaar,
+  Malwares,
+  NVD,
+  OCCRP,
+  ONYPHE,
+  OTX,
+  OpenTIP,
+  Pulsedive,
+  Radar,
+  RiskIQ,
+  Robtex,
+  Scumware,
+  SecurityTrails,
+  Shodan,
+  Sploitus,
+  SpyOnWeb,
+  TIP,
+  Talos,
+  ThreatConnect,
+  ThreatMiner,
+  Triage,
+  URLScan,
+  URLVoid,
+  URLhaus,
+  VMRay,
+  ViewDNS,
+  VirusTotal,
+  Vulmon,
+  VulncodeDB,
+  XForceExchange,
+  ZoomEye,
+} from "./index";
+
 export { AbuseIPDB } from "./abuseipdb";
 export { AnyRun } from "./anyrun";
 export { APKLab } from "./apklab";
@@ -72,4 +143,71 @@ export { XForceExchange } from "./xforceExchange";
 export { ZoomEye } from "./zoomeye";
 export { All } from "./all";
 
-export { Searchers } from "./searchers";
+export const Searchers: Searcher[] = [
+  new AbuseIPDB(),
+  new AnyRun(),
+  new APKLab(),
+  new ArchiveOrg(),
+  new ArchiveToday(),
+  new BGPView(),
+  new BinaryEdge(),
+  new BitcoinAbuse(),
+  new BitcoinWhosWho(),
+  new BlockChain(),
+  new Blockchair(),
+  new BlockCypher(),
+  new Censys(),
+  new CheckPhish(),
+  new Crtsh(),
+  new DNSlytics(),
+  new DomainTools(),
+  new DomainWatch(),
+  new EmailRep(),
+  new ExploitDatabase(),
+  new FileScan(),
+  new FortiGuard(),
+  new GoogleSafeBrowsing(),
+  new GreyNoise(),
+  new Hashdd(),
+  new HurricaneElectric(),
+  new HybridAnalysis(),
+  new InQuest(),
+  new IntelligenceX(),
+  new Intezer(),
+  new IPinfo(),
+  new IPIP(),
+  new JoeSandbox(),
+  new Malshare(),
+  new Maltiverse(),
+  new MalwareBazaar(),
+  new Malwares(),
+  new NVD(),
+  new OCCRP(),
+  new ONYPHE(),
+  new OpenTIP(),
+  new OTX(),
+  new Pulsedive(),
+  new Radar(),
+  new RiskIQ(),
+  new Robtex(),
+  new Scumware(),
+  new SecurityTrails(),
+  new Shodan(),
+  new Sploitus(),
+  new SpyOnWeb(),
+  new Talos(),
+  new ThreatConnect(),
+  new ThreatMiner(),
+  new TIP(),
+  new Triage(),
+  new URLhaus(),
+  new URLScan(),
+  new URLVoid(),
+  new ViewDNS(),
+  new VirusTotal(),
+  new VMRay(),
+  new Vulmon(),
+  new VulncodeDB(),
+  new XForceExchange(),
+  new ZoomEye(),
+];
